Bring Mac XCode project in sync with next preview
Also rearchitected how the common dependencies work to statically reference them. (Ogg, Vorbis, and SDL)
These changes are in preparation for iOS at some point in the future since it's preferred to statically link there instead of using Frameworks.
This also makes it easier for other people to build the code, because I no longer need to bundle pre-built copies of those dependencies.